requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
boto3==1.34.0
cloudinary==1.36.0
reportlab==4.0.4
schedule==1.2.0
paramiko==3.3.1
Pillow==10.0.1
